+++ R/oauth-flow-device.R
@@ -13,11 +13,17 @@
 #' @inheritParams req_oauth_auth_code
 #' @returns A modified HTTP [request].
 #' @examples
-#' client <- oauth_client("example", "https://example.com/get_token")
-#' req <- request("https://example.com")
+#' req_auth_github <- function(req) {
+#'   req_oauth_device(
+#'     req,
+#'     client = example_github_client(),
+#'     auth_url = "https://github.com/login/device/code"
+#'   )
+#' }
 #'
-#' req %>% req_oauth_device(client)
-req_oauth_device <- function(req, client,
+#' request("https://api.github.com/user") %>%
+#'   req_auth_github()
+req_oauth_device <- function(req, client, auth_url,
                              cache_disk = FALSE,
                              cache_key = NULL,
                              scope = NULL,
@@ -26,6 +32,7 @@ req_oauth_device <- function(req, client,
 
   params <- list(
     client = client,
+    auth_url = auth_url,
     scope = scope,
     auth_params = auth_params,
     token_params = token_params
@@ -98,18 +105,25 @@ oauth_flow_device <- function(client,
 # Device authorization request and response
 # https://datatracker.ietf.org/doc/html/rfc8628#section-3.1
 # https://datatracker.ietf.org/doc/html/rfc8628#section-3.2
-oauth_flow_device_request <- function(client, auth_url, scope, auth_params) {
+oauth_flow_device_request <- function(client,
+                                      auth_url,
+                                      scope,
+                                      auth_params,
+                                      error_call = caller_env()) {
   req <- request(auth_url)
   req <- req_body_form(req, scope = scope, !!!auth_params)
   req <- oauth_client_req_auth(req, client)
   req <- req_headers(req, Accept = "application/json")
 
-  oauth_flow_fetch(req)
+  oauth_flow_fetch(req, "auth_url", error_call = error_call)
 }
 
 # Device Access Token Request
 # https://datatracker.ietf.org/doc/html/rfc8628#section-3.4
-oauth_flow_device_poll <- function(client, request, token_params) {
+oauth_flow_device_poll <- function(client,
+                                   request,
+                                   token_params,
+                                   error_call = caller_env()) {
   cli::cli_progress_step("Waiting for response from server", spinner = TRUE)
 
   delay <- request$interval %||% 5
@@ -127,7 +141,8 @@ oauth_flow_device_poll <- function(client, request, token_params) {
         token <- oauth_client_get_token(client,
           grant_type = "urn:ietf:params:oauth:grant-type:device_code",
           device_code = request$device_code,
-          !!!token_params
+          !!!token_params,
+          error_call = error_call
         )
         break
       },